+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed lint staged command in order to check typescript errors on commits
+- Fixed Loading component in order to work with android 7 devices
 
 ## [1.0.3] - 2023-08-24